hack: Always create the metering PrometheusRules custom resource in the monitoring namespace
Previously, I had the metadata.namespace set to the `openshift-monitoring` namespace for both the monitoring-related manifests (cluster-monitoring-operator config and metering PrometheusRules), which was working for me locally. When I attempted to create a more generalistic approach to creating the custom PrometheusRules custom resource in a non-openshift-prefixed namespace, it appears those custom rule definitions weren't being created and/or respected. After lightly diving into this, it appears that Prometheus isn't configured to watch any non-openshift-prefixed namespaces, or at least rule definitions created outside of those namespaces aren't a supported functionality in this monitoring stack, which results in Prometheus not adding these rules to the list of available metrics.

Hardcoding the metadata.namespace to the openshift-monitoring namespace ended up fixing this, despite the following log message in the Prometheus Operator logs:
```log
│ E0921 14:16:51.956204       1 reflector.go:280] github.com/coreos/prometheus-operator/pkg/thanos/operator.go:319: Failed to watch *v1.PrometheusRule: expected resource version to have 38 parts to match the number of ListerWatchers                                                                                     ││ E0921 14:17:27.011331       1 reflector.go:280] github.com/coreos/prometheus-operator/pkg/prometheus/operator.go:480: Failed to watch *v1.PrometheusRule: expected resource version to have 38 parts to match the number of ListerWatchers
```