manifests/presto: Fix the certificate CA bundle that gets mounted in Presto
In Openshift 4.5+, there's no controller responsible for mounting the service serving CA bundle into containers for us, which leads to the situation where we are required to create an empty ConfigMap containing a specific label or annotation such that a controller responsible for injecting that CA bundle into the ConfigMap data understands, and mounting that ConfigMap manually as a volume/volumeMount at the container-level. Previously, I was running into authentication issues where Presto and Prometheus were unable to communicate with each other, despite a valid bearer token file configuration being specified. After some digging in that coordinator's container, it appears that I neglected to use the correct ConfigMap metadata configuration. As a result, the `config.openshift.io/inject-trusted-cabundle` label was used, which mounts as series of self-signed certificates, which was causing those authentication issues. If we switch to using the correct metadata configuration, i.e. specifying this `service.beta.openshift.io/inject-cabundle: "true"` annotation, then the correct CA bundle gets mounted and we can add that to the default Java truststore.